Here are the results I got by running it locally. master vs https://github.com/root-project/root/pull/7651:

```
# master
10: BM_RDataFrame_JitLargeGraphs/100                                  839628 us       832963 us            1
10: BM_RDataFrame_JitLargeGraphs/1000                                3170937 us      3165917 us            1
10: BM_RDataFrame_JitLargeGraphs/10000                              38968494 us     38905225 us            1

# optimized
10: BM_RDataFrame_JitLargeGraphs/100                                  683157 us       677816 us            1
10: BM_RDataFrame_JitLargeGraphs/1000                                1284963 us      1282854 us            1
10: BM_RDataFrame_JitLargeGraphs/10000                              11771816 us     11750960 us            1
```