@eguiraud 2021-03-24
Thanks @stwunsch ! How much of the test time is spent in the event loop? (i.e. can we conflate test time with jitting time or are there caveats?)