@stwunsch 2021-03-25
@eguiraud It's 100% spent in jitting, the event loop is very fast because the depth of the graph is just 3. See here with logging:

```
10: Info in <[ROOT.RDF] Info /work/wunsch/root_jit/tree/dataframe/src/RLoopManager.cxx:629 in void ROOT::Detail::RDF::RLoopManager::Jit()>: Just-in-time compilation phase completed in 0.853629 seconds.
10: Info in <[ROOT.RDF] Info /work/wunsch/root_jit/tree/dataframe/src/RLoopManager.cxx:678 in void ROOT::Detail::RDF::RLoopManager::Run()>: Finished event loop number 0 (0s CPU, 0.000236034s elapsed).

10: Info in <[ROOT.RDF] Info /work/wunsch/root_jit/tree/dataframe/src/RLoopManager.cxx:629 in void ROOT::Detail::RDF::RLoopManager::Jit()>: Just-in-time compilation phase completed in 1.321165 seconds.
10: Info in <[ROOT.RDF] Info /work/wunsch/root_jit/tree/dataframe/src/RLoopManager.cxx:678 in void ROOT::Detail::RDF::RLoopManager::Run()>: Finished event loop number 0 (0.01s CPU, 0.00238919s elapsed).

10: Info in <[ROOT.RDF] Info /work/wunsch/root_jit/tree/dataframe/src/RLoopManager.cxx:629 in void ROOT::Detail::RDF::RLoopManager::Jit()>: Just-in-time compilation phase completed in 12.077101 seconds.
10: Info in <[ROOT.RDF] Info /work/wunsch/root_jit/tree/dataframe/src/RLoopManager.cxx:678 in void ROOT::Detail::RDF::RLoopManager::Run()>: Finished event loop number 0 (0.05s CPU, 0.043184s elapsed).
```